store: do not form connections to connections, closes #28
this fixes parent -> child -> parent object passing without any weird breakable-by-opcache anti-features that cause big performance losses.

pthreads_zend_object_t now stores a pointer to the original pthreads_zend_object_t to which it is a connection, assuming that it is one. This allows subsequent connections to connect directly to the original object instead of connecting to a connection which might live less long than the original object.