Always init static_variables_ptr mapped pointer
I don't know how many crashes this was responsible for; probably not many since it would only show up on non-immutable functions with statics, but it depends on a variety of factors that I didn't take the time to investigate.
What I do know is that zend expects this to _always_ be initialized for mutable functions even if there are no static vars.