Allow strings to be directly copied between threads in some cases (#93)
closes #71

This works by allowing threads direct access to the original thread's string for as long as it is cached in the origin thread's ThreadedBase connection.

As long as the connection lives long enough for the reading thread(s) to dereference the string in question (which, thanks to 4ddf79e0888ad43211922fce1d29c5bc3b481837, will be the case in all common cases, including dead threads and completed worker tasks), the extra malloc and string copy is avoided on the writer thread, which significantly improves performance in synthetic benchmarks.

If the connection is destroyed, it will create persistent copies of any cached strings during free_obj, and the old double-copy method will be used to enable the string to be accessed. However, this is rarely needed.

The caveat to this is that pthreads_store_sync_local_properties() will do work more often when strings are used, but I don't think this is a big concern. For most cases, the property table should be small enough for this to not be a problem anyway, and for the large cases, we need to implement dedicated queue data structures anyway. Profiling anyway suggested that the overhead of zend_hash_internal_pointer_reset_ex() was several orders of magnitude bigger a problem anyway (see #42).