Begin making closures more sane (#97) (#101)
Previously, copied closures would be silently altered by pthreads if they contained stuff pthreads didn't like. In addition, no effort whatsoever was made to ensure that static variables behaved in any kind of sane manner.

This change:
- prohibits non-static closures with a non-thread-safe $this bound (previously they would be copied with a null $this, leading to random crashes when called)
- restores $this onto target threads if $this is a thread-safe object (previously would be NULLed)
- disallows the use of static variables inside copied closures (since they can't be synchronized, and may contain non-thread-safe values after being called)
- disallows the use of use-by-reference (for obvious reasons)

We do also need to restrict the stuff that can be use()d into a closure, but that's a more complex endeavour, so this will have to be enough for now.