First look at readonly property support
Before this, readonly was silently ignored, so properties would be
treated the same as normal properties.

This change enforces write-once and no-unset policies for readonly
properties. It also permits lockless reads of locally cached property
values of readonly properties.

Currently this only covers stuff that is kept in the local cache, such
as thread-safe objects, closures, sockets and strings. The cache could
be extended to more stuff to permit this optimisation to apply to other
types of properties, but that's a task for another time.