Thread::start() now requires options to be specified
Since there's no one-size-fits-all for this, we require the user to specify it.

For single-file scripts, it'll usually be necessary to use INHERIT_ALL, since you'll probably want to use classes, functions and constants in the same file that can't be autoloaded.
For production applications, which usually follow a one-class-per-file rule, INHERIT_NONE or INHERIT_INI is preferable, since these have significantly less overhead and use much less memory.